+++ src/main/scala/com/jcdecaux/setl/internal/CanDrop.scala
@@ -2,6 +2,9 @@ package com.jcdecaux.setl.internal
 
 trait CanDrop {
 
+  /**
+   * Drop the entire table.
+   */
   def drop(): Unit
 
 }