+++ src/main/scala/com/jcdecaux/setl/storage/repository/streaming/package.scala
@@ -1,0 +1,13 @@
+package com.jcdecaux.setl.storage.repository
+
+import scala.reflect.runtime.universe.TypeTag
+
+package object streaming {
+
+  implicit class SparkRepositoryStreamingImplicits[T: TypeTag](sparkRepository: SparkRepository[T]) {
+
+    def toStreamingRepository: StreamingRepository[T] = StreamingRepository(sparkRepository)
+
+  }
+
+}