+++ src/main/scala/com/jcdecaux/setl/config/DynamoDBConnectorConf.scala
@@ -1,5 +1,7 @@
 package com.jcdecaux.setl.config
 
+import org.apache.spark.sql.SaveMode
+
 class DynamoDBConnectorConf extends ConnectorConf {
 
   import DynamoDBConnectorConf._
@@ -12,9 +14,15 @@ class DynamoDBConnectorConf extends ConnectorConf {
 
   def getReadPartitions: Option[String] = get(Reader.READ_PARTITIONS)
 
+  def getSaveMode: SaveMode = SaveMode.valueOf(get("saveMode", "ErrorIfExists"))
+
   override def getReaderConf: Map[String, String] = {
     import scala.collection.JavaConverters._
-    settings.asScala.toMap - Writer.WRITE_BATCH_SIZE - Writer.UPDATE - TABLE
+    settings.asScala.toMap -
+      Writer.WRITE_BATCH_SIZE -
+      Writer.UPDATE -
+      TABLE -
+      Writer.SAVE_MODE
   }
 
   override def getWriterConf: Map[String, String] = {
@@ -25,7 +33,9 @@ class DynamoDBConnectorConf extends ConnectorConf {
       Reader.DEFAULT_PARALLELISM -
       Reader.STRONGLY_CONSISTENT_READS -
       Reader.BYTES_PER_RCU -
-      Reader.FILTER_PUSHDOWN - TABLE
+      Reader.FILTER_PUSHDOWN -
+      TABLE -
+      Writer.SAVE_MODE
   }
 
   def getRegion: Option[String] = get(REGION)
@@ -46,6 +56,7 @@ object DynamoDBConnectorConf {
   object Writer {
     val WRITE_BATCH_SIZE = "writeBatchSize"
     val UPDATE = "update"
+    val SAVE_MODE = "saveMode"
   }
 
   val REGION: String = "region"